feat(registry): gitlab browse projects without extension
